add Eureka
Eureka is a simple add-on for Valkyrien Skies that fits with vanilla
aesthetic and is easy and fun to use.

Build and design your own ship out of ordinary (or modded!) Minecraft
blocks, place a Ship Helm, shift + right-click, assemble, and watch your
creation suddenly infuse with physics. No longer are your Minecraft
builds bound to remain forever static, sitting in the same place for all
eternity.

Signed-off-by: Troy <troymitchell988@gmail.com>