add ClockworkAdditions
Clockwork Additions is a QoL mod that adds several features to Clockwork
and Valkyrien Skies

Blocks and Features:

The Pedal Block produces SU when a player is sitting on it and
attempting to walk forward. It produces 320 SU at 32 RPM normally, and
640 SU at 64 RPM when sprinting. It can also be reversed by attempting
to walk backwards.

The Handlebar is controlled by sitting in front of it and using the A
and D keys. It will rotate 45 degrees in each direction with 128 SU.

The Constant Velocity Joint is flexible shaft that works between ships.
It has a maximum range of 3 blocks.

The Kinetic Flap Bearing is a Clockwork Flap Bearing that rotates when
given clockwise or counterclockwise rotation, instead of being activated
with redstone.

The Copycat Wing and Flap work as normal wings and flaps, but can also
have their textures changed by right-clicking them with a block.

Signed-off-by: Troy Mitchell <troymitchell988@gmail.com>