add CC:tweaked and CC:vs
CC: Tweaked is a mod for Minecraft which adds programmable computers,
turtles and more to the game. A fork of the much-beloved ComputerCraft,
it continues its legacy with improved performance and stability, along
with a wealth of new features.

CC:VS is a minimalist integration addon for Valkyrien Skies 2 to
let ComputerCraft computers gather more information and (in the case of
Command Computers) more control over Ships

There are only two features in this addon, both of which are APIs for
computers placed on Ships

ShipAPI - a new API that allows computers placed on Ships greater access
to Ship information! It will appear listed as both "ship" and as the
Ship's name at the time of startup!

ExtendedShipAPI - This API is an extension of ShipAPI specifically for
Command Computers and has inherited the original methods as well as new
methods for direct Ship control! It will also appear listed as both
"ship" and as the Ship's name at the time of startup!

There are also, as of 0.0.8, a few ingame utility programs for creating
custom PID controllers thanks to 19PHOBOSS98 allowing me to add them
from his Github Repo which is under the MIT License!

Signed-off-by: Troy Mitchell <troymitchell988@gmail.com>